backend/requirements.txt
Django==5.2.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
mongoengine==0.28.2
django-cors-headers==4.4.0
google-auth==2.35.0
requests==2.32.5
python-dotenv==1.0.1
